util: move zio into libutil
Split standalone zio.[ch] from the kz and forkzio sources since it doesn't
have dependency on kvs or other broker modules. Place zio into libutil for
easy inclusion in other sources, and also to simplify linking.
