+++ src/bindings/lua/flux-lua.c
@@ -39,7 +39,7 @@
 
 #include "src/common/libutil/jsonutil.h"
 #include "src/modules/libmrpc/mrpc.h"
-#include "src/common/libzio/zio.h"
+#include "src/common/libutil/zio.h"
 #include "src/common/libzio/kz.h"
 
 #include "json-lua.h"
