+++ src/common/libutil/Makefile.am
@@ -42,7 +42,9 @@ libutil_la_SOURCES = \
 	ev_zlist.c \
 	ev_zlist.h \
 	coproc.c \
-	coproc.h
+	coproc.h \
+	zio.c \
+	zio.h
 
 EXTRA_DIST = veb_mach.c
