+++ src/common/libzio/forkzio.c
@@ -37,8 +37,8 @@
 
 #include "src/common/libutil/log.h"
 #include "src/common/libutil/xzmalloc.h"
+#include "src/common/libutil/zio.h"
 
-#include "zio.h"
 #include "forkzio.h"
 
 struct forkzio_handle_struct {
